provide more logging
